backend/flask-image-upload-2/requirements.txt
Flask==2.0.3
numpy==1.21.2
Pillow==8.4.0
scikit-learn==0.24.2
tensorflow==2.6.0